uci.cpp fix race conditions introduced in my previous commit
I'm still trying to understand the code, but it appeares that simply
undoing a move can affect the states and cause issues when navigating
back and forth through a game. Resetting the whole position seems more
robust.

Introduced this method to set the position each time position, move,
back or reset are called:

  void updatePosition(Position& pos)

These variable are moved to global scope (some were in loop()):
  fen, sfen, extStates, moveStack, sanMoves

- states was renamed extStates to avoid confusion with local variable
  states.
- moveStack is the list of moves to apply after setting the fen.
- sanMoves is the unprocessed list of move strings entered with the move
  command.